CalculationLibrary/Calculator.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace CalculationLibrary
{
    public class Calculator
    {
        public double Add(double x, double y)
        {
            return x + y;
        }

        public double Subtract(double x, double y)
        {
            return x - y;
        }

        public double Multiply(double x, double y)
        {
            return x * y; 
        }

        public double Divide(double x, double y)
        {
            double output = 0;
            if (y != 0)
            {
                output = x / y;
            }
            return output;
        }
    }
}
